Introduce utility class `DeltaLakeDomains` in Delta Lake
Extracted `partitionMatchesPredicate` from `DeltaLakeSplitManager` into a new utility class `DeltaLakeDomains`
